+++ CompetitionManagementSystem/src/main/java/com/qiaose/competitionmanagementsystem/entity/User.java
@@ -47,7 +47,7 @@ public class User implements Serializable {
     * 积分
     */
     @TableField(exist = false)
-    private Integer userIntegral;
+    private Double userIntegral;
 
     /**
     * 创建时间
